refactor: 애플 OAuth 이름 추출 로직을 공통 resolver로 통합
