+++ src/main/java/gg/agit/konect/infrastructure/oauth/AppleOAuthServiceImpl.java
@@ -1,7 +1,6 @@
 package gg.agit.konect.infrastructure.oauth;
 
 import java.util.Optional;
-import java.util.Map;
 
 import org.springframework.security.oauth2.client.oidc.userinfo.OidcUserRequest;
 import org.springframework.security.oauth2.client.oidc.userinfo.OidcUserService;
@@ -16,6 +15,7 @@ import gg.agit.konect.domain.user.model.UnRegisteredUser;
 import gg.agit.konect.domain.user.model.User;
 import gg.agit.konect.domain.user.repository.UnRegisteredUserRepository;
 import gg.agit.konect.domain.user.repository.UserRepository;
+import gg.agit.konect.global.auth.oauth.AppleOAuthNameResolver;
 import gg.agit.konect.global.code.ApiResponseCode;
 import gg.agit.konect.global.exception.CustomException;
 import lombok.RequiredArgsConstructor;
@@ -27,6 +27,7 @@ public class AppleOAuthServiceImpl extends OidcUserService {
 
     private final UserRepository userRepository;
     private final UnRegisteredUserRepository unRegisteredUserRepository;
+    private final AppleOAuthNameResolver appleOAuthNameResolver;
 
     @Transactional
     @Override
@@ -35,7 +36,7 @@ public class AppleOAuthServiceImpl extends OidcUserService {
 
         String email = oidcUser.getAttribute("email");
         String providerId = oidcUser.getSubject();
-        String name = extractName(oidcUser);
+        String name = appleOAuthNameResolver.resolve(oidcUser.getAttributes());
 
         String registrationId = userRequest.getClientRegistration().getRegistrationId().toUpperCase();
         Provider provider = Provider.valueOf(registrationId);
@@ -70,56 +71,4 @@ public class AppleOAuthServiceImpl extends OidcUserService {
         return oidcUser;
     }
 
-    private String extractName(OidcUser oidcUser) {
-        String name = oidcUser.getAttribute("name");
-
-        if (StringUtils.hasText(name)) {
-            return name;
-        }
-
-        String givenName = oidcUser.getAttribute("given_name");
-        String familyName = oidcUser.getAttribute("family_name");
-
-        if (StringUtils.hasText(givenName) && StringUtils.hasText(familyName)) {
-            return familyName + givenName;
-        }
-
-        if (StringUtils.hasText(givenName)) {
-            return givenName;
-        }
-
-        if (StringUtils.hasText(familyName)) {
-            return familyName;
-        }
-
-        Object rawName = oidcUser.getAttributes().get("name");
-        if (!(rawName instanceof Map<?, ?> nameMap)) {
-            return null;
-        }
-
-        String firstName = asText(nameMap.get("firstName"));
-        String lastName = asText(nameMap.get("lastName"));
-
-        if (StringUtils.hasText(firstName) && StringUtils.hasText(lastName)) {
-            return lastName + firstName;
-        }
-
-        if (StringUtils.hasText(firstName)) {
-            return firstName;
-        }
-
-        if (StringUtils.hasText(lastName)) {
-            return lastName;
-        }
-
-        return null;
-    }
-
-    private String asText(Object value) {
-        if (value instanceof String text && StringUtils.hasText(text)) {
-            return text;
-        }
-
-        return null;
-    }
 }
